Add application data cleanup logic (#116)
Case 1: Invalid application name
```
[root@temp ai-services]# ./bin/ai-services application delete temp/
Error: invalid application name: temp/
Usage:
  ai-services application delete [name] [flags]

Flags:
  -h, --help           help for delete
      --skip-cleanup   Skip deleting application data (default=false)
```
```
[root@temp ai-services]# ./bin/ai-services application delete ../temp
Error: invalid application name: ../temp
Usage:
  ai-services application delete [name] [flags]

Flags:
  -h, --help           help for delete
      --skip-cleanup   Skip deleting application data (default=false)
```
```
[root@temp ai-services]# ./bin/ai-services application delete ..
Error: invalid application name: ..
Usage:
  ai-services application delete [name] [flags]

Flags:
  -h, --help           help for delete
      --skip-cleanup   Skip deleting application data (default=false)
```
---
Case: 2 Dir doesnt exists
```
[root@temp ai-services]# ls /var/lib/ai-services/
applications  models
[root@temp ai-services]# ls /var/lib/ai-services/applications/
```
```
[root@temp ai-services]# ./bin/ai-services application delete temp 
Found 3 pods for given applicationName: temp.
Below are the list of pods to be deleted
	-> temp--pod-1
	-> temp--pod-2
	-> temp--pod-3
Are you sure you want to delete above pods?  true
Proceeding with deletion...
Deleting the pod: temp--pod-1
Successfully removed the pod: temp--pod-1
Deleting the pod: temp--pod-2
Successfully removed the pod: temp--pod-2
Deleting the pod: temp--pod-3
Successfully removed the pod: temp--pod-3
```
---
Case 3: `--skip-cleanup` flag is provided
```
[root@temp ai-services]# ls /var/lib/ai-services/applications/
temp
```
```
[root@temp ai-services]# ./bin/ai-services application delete temp --skip-cleanup
Found 3 pods for given applicationName: temp.
Below are the list of pods to be deleted
	-> temp--pod-1
	-> temp--pod-3
	-> temp--pod-2
Are you sure you want to delete above pods?  true
Proceeding with deletion...
Deleting the pod: temp--pod-1
Successfully removed the pod: temp--pod-1
Deleting the pod: temp--pod-3
Successfully removed the pod: temp--pod-3
Deleting the pod: temp--pod-2
Successfully removed the pod: temp--pod-2
```
```
[root@temp ai-services]# ls /var/lib/ai-services/applications/
temp
```
---
Case 4: Again trigger delete, but without `--skip-cleanup`
```
[root@temp ai-services]# ls /var/lib/ai-services/applications/
temp
```
```
[root@temp ai-services]# ./bin/ai-services application delete temp
No pods found with given application: temp
Cleaning up application data
Application data cleaned up successfully
```
---
Case 5: AppDir exists, and deletion is triggered
```
[root@temp ai-services]# ls /var/lib/ai-services/applications/
temp
```
```
[root@temp ai-services]# ./bin/ai-services application delete temp
Found 3 pods for given applicationName: temp.
Below are the list of pods to be deleted
	-> temp--pod-1
	-> temp--pod-2
	-> temp--pod-3
Are you sure you want to delete above pods?  true
Proceeding with deletion...
Deleting the pod: temp--pod-1
Successfully removed the pod: temp--pod-1
Deleting the pod: temp--pod-2
Successfully removed the pod: temp--pod-2
Deleting the pod: temp--pod-3
Successfully removed the pod: temp--pod-3
Cleaning up application data
Application data cleaned up successfully
```

+++ ai-services/cmd/ai-services/cmd/application/delete.go
@@ -2,14 +2,21 @@ package application
 
 import (
 	"fmt"
+	"os"
+	"path/filepath"
 	"strings"
 
 	"github.com/containers/podman/v5/pkg/domain/entities/types"
 	"github.com/spf13/cobra"
 
+	"github.com/project-ai-services/ai-services/internal/pkg/constants"
 	"github.com/project-ai-services/ai-services/internal/pkg/logger"
 	"github.com/project-ai-services/ai-services/internal/pkg/runtime/podman"
 	"github.com/project-ai-services/ai-services/internal/pkg/utils"
+)
+
+var (
+	skipCleanup bool
 )
 
 var deleteCmd = &cobra.Command{
@@ -20,6 +27,11 @@ var deleteCmd = &cobra.Command{
 Arguments
   [name]: Application name (required)`,
 	Args: cobra.ExactArgs(1),
+	PreRunE: func(cmd *cobra.Command, args []string) error {
+		appName := args[0]
+
+		return utils.VerifyAppName(appName)
+	},
 	RunE: func(cmd *cobra.Command, args []string) error {
 		applicationName := args[0]
 
@@ -42,7 +54,19 @@ Arguments
 	},
 }
 
+func init() {
+	deleteCmd.Flags().BoolVar(&skipCleanup, "skip-cleanup", false, "Skip deleting application data (default=false)")
+}
+
 func deleteApplication(client *podman.PodmanClient, appName string) error {
+
+	appDir := filepath.Join(constants.ApplicationsPath, filepath.Base(appName))
+	appExists := false
+
+	if _, err := os.Stat(appDir); err == nil {
+		appExists = true
+	}
+
 	resp, err := client.ListPods(map[string][]string{
 		"label": {fmt.Sprintf("ai-services.io/application=%s", appName)},
 	})
@@ -57,44 +81,65 @@ func deleteApplication(client *podman.PodmanClient, appName string) error {
 		pods = val
 	}
 
-	if len(pods) == 0 {
-		logger.Infof("No pods found with given application: %s\n", appName)
+	podsExists := len(pods) != 0
+
+	if podsExists {
+		logger.Infof("Found %d pods for given applicationName: %s.\n", len(pods), appName)
+		logger.Infoln("Below are the list of pods to be deleted")
+		for _, pod := range pods {
+			logger.Infof("\t-> %s\n", pod.Name)
+		}
+	} else {
+		logger.Infof("No pods found for application: %s\n", appName)
+	}
+
+	var confirmActionPrompt string
+	if podsExists && appExists && !skipCleanup {
+		confirmActionPrompt = "Are you sure you want to delete the above pods and application data? "
+	} else if podsExists {
+		confirmActionPrompt = "Are you sure you want to delete the above pods? "
+	} else if appExists && !skipCleanup {
+		confirmActionPrompt = "Are you sure you want to delete the application data? "
+	} else {
+		logger.Infof("Application %s does not exist", appName)
 		return nil
 	}
 
-	logger.Infof("Found %d pods for given applicationName: %s.\n", len(pods), appName)
-	logger.Infoln("Below are the list of pods to be deleted")
-	for _, pod := range pods {
-		logger.Infof("\t-> %s\n", pod.Name)
-	}
-
-	confirmDelete, err := utils.ConfirmAction("Are you sure you want to delete above pods? ")
+	confirmDelete, err := utils.ConfirmAction(confirmActionPrompt)
 	if err != nil {
 		return fmt.Errorf("failed to take user input: %w", err)
 	}
 
 	if !confirmDelete {
-		logger.Infof("Skipping the deletion of pods")
+		logger.Infoln("Deletion cancelled")
 		return nil
 	}
 
-	logger.Infof("Proceeding with deletion...\n")
+	logger.Infoln("Proceeding with deletion...")
 
-	// Loop over each of the pods and call delete
 	var errors []string
-	for _, pod := range pods {
-		logger.Infof("Deleting the pod: %s\n", pod.Name)
-		if err := client.DeletePod(pod.Id, utils.BoolPtr(true)); err != nil {
-			errMsg := fmt.Sprintf("%s: %v", pod.Name, err)
-			errors = append(errors, errMsg)
-			continue
+	if podsExists {
+		for _, pod := range pods {
+			logger.Infof("Deleting pod: %s\n", pod.Name)
+			if err := client.DeletePod(pod.Id, utils.BoolPtr(true)); err != nil {
+				errors = append(errors, fmt.Sprintf("pod %s: %v", pod.Name, err))
+				continue
+			}
+			logger.Infof("Successfully removed pod: %s\n", pod.Name)
 		}
-		logger.Infof("Successfully removed the pod: %s\n", pod.Name)
 	}
 
 	// Aggregate errors at the end
 	if len(errors) > 0 {
 		return fmt.Errorf("failed to remove pods: \n%s", strings.Join(errors, "\n"))
+	}
+
+	if appExists && !skipCleanup {
+		logger.Infoln("Cleaning up application data")
+		if err := os.RemoveAll(appDir); err != nil {
+			return fmt.Errorf("failed to delete application data: %w", err)
+		}
+		logger.Infoln("Application data cleaned up successfully")
 	}
 
 	return nil
